[editor] Add editor/shared/hooks. Move usePending into editor/shared/hooks.
