+++ app/src/editor/canvas/components/CanvasController/useCanvasDuplicateCallback.jsx
@@ -1,10 +1,11 @@
 import { useContext, useCallback } from 'react'
-import * as RouterContext from '$editor/shared/components/RouterContext'
+
 import useIsMountedRef from '$shared/utils/useIsMountedRef'
+import * as RouterContext from '$editor/shared/components/RouterContext'
+import usePending from '$editor/shared/hooks/usePending'
 
 import links from '../../../../links'
 import * as services from '../../services'
-import usePending from './usePending'
 
 export default function useCanvasDuplicateCallback() {
     const { history } = useContext(RouterContext.Context)
